Refactor: Split Cardano.CLI.Compatible.Transaction into separate modules
