+++ cardano-cli/src/Cardano/CLI/Compatible/Transaction/Commands.hs
@@ -1,0 +1,44 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE ExistentialQuantification #-}
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+
+module Cardano.CLI.Compatible.Transaction.Commands
+  ( CompatibleTransactionCmds (..)
+  , renderCompatibleTransactionCmd
+  )
+where
+
+import           Cardano.Api
+import           Cardano.Api.Ledger hiding (TxIn, VotingProcedures)
+
+import           Cardano.CLI.EraBased.Script.Certificate.Types
+import           Cardano.CLI.EraBased.Script.Proposal.Types
+import           Cardano.CLI.EraBased.Script.Vote.Types (CliVoteScriptRequirements)
+import           Cardano.CLI.Types.Common
+import           Cardano.CLI.Types.Governance
+
+import           Data.Text (Text)
+
+-- TODO: After QA confirmms that the new compatibility commands meet their needs
+-- we can remove all remaining legacy commands. We can also remove/move the exising
+-- byron era commands under the new compatiblilty commands.
+data CompatibleTransactionCmds era
+  = CreateCompatibleSignedTransaction
+      (ShelleyBasedEra era)
+      [TxIn]
+      [TxOutAnyEra]
+      !(Maybe (Featured ShelleyToBabbageEra era (Maybe UpdateProposalFile)))
+      !(Maybe (Featured ConwayEraOnwards era [(ProposalFile In, Maybe CliProposalScriptRequirements)]))
+      ![(VoteFile In, Maybe CliVoteScriptRequirements)]
+      [WitnessSigningData]
+      -- ^ Signing keys
+      (Maybe NetworkId)
+      !Coin
+      -- ^ Tx fee
+      ![(CertificateFile, Maybe CliCertificateScriptRequirements)]
+      -- ^ stake registering certs
+      !(File () Out)
+
+renderCompatibleTransactionCmd :: CompatibleTransactionCmds era -> Text
+renderCompatibleTransactionCmd _ = ""
